requirements.txt
djangorestframework>=3.14.0
drf-nested-routers>=0.93.4
django-filter>=23.3
Pillow>=10.0.0
python-dateutil>=2.8.2
